only consider data inside the microenvironment
Previously, we would get distances that went outside of the tumor
microenvironment. This is probably because of how the shapely buffer
function works. It seems that the distance of buffering not the same at
all points in the polygon, and some points might be buffered further. So
we filter out any points that are outside of the environment.