only use exterior tumor boundaries
This commit re-implements the tumor boundary as a MultiLineString
object. The marker-positive multiline string contains the lines
belonging to marker-positive blocks, and similar for marker-negative.

Additionally, this commit fixes the tumor microenvironment calculation.
Now, it is a dilation of the tumor exterior, not of the entire tumor
polygon. This means that now, we should not be selecting cells inside
the tumor region (only within the microenvironment.