fix bug in right_x and bottom_y!!
Now that this bug is fixed, the correct patches are chosen. Now the
outputs make more sense.

This commit also adds support for polygons and multipolygons for some of
the functions that get exteriors of shapely objects.

In addition, this commit adds a method to plot edges of tumor regions
and the cells on an image.