fix a type error in point output instantiation
we will probably have to implement the normalization code in a different
way. commit
https://github.com/kaczmarj/tumor-microenvironment-analysis/commit/b7510a6ce0880eba17ff8542a68601ceacc9e850#diff-f815915a72feccdb45a267421bea4d43b92c9bf415376dc9e374c89ccf0a1623
introduces a lot of code duplication. we can avoid this by instead
getting the distance just from the centroid and not each lattice point.
the function that iterates over lattice points would have to be
modified, but many other functions can be left alone.